+++ src/main/java/com/scs/voxlib/chunk/VoxTransformChunk.java
@@ -1,4 +1,7 @@
-package com.scs.voxlib;
+package com.scs.voxlib.chunk;
+
+import com.scs.voxlib.GridPoint3;
+import com.scs.voxlib.StreamUtils;
 
 import java.io.IOException;
 import java.io.InputStream;
