+++ src/main/java/com/scs/voxlib/mat/VoxMaterial.java
@@ -1,4 +1,4 @@
-package com.scs.voxlib;
+package com.scs.voxlib.mat;
 
 import java.util.HashMap;
 import java.util.Objects;
