+++ src/main/java/com/scs/voxlib/mat/VoxOldMaterialType.java
@@ -1,4 +1,4 @@
-package com.scs.voxlib;
+package com.scs.voxlib.mat;
 
 import java.util.Optional;
